use vec for URL and some rafactoring
