+++ src/components/worker.rs
@@ -14,7 +14,7 @@ use log;
 
 use self::{request::WorkerRequest, result::WorkerResult};
 
-use super::http::{client::HttpClient, HttpResult};
+use crate::components::http::{client::HttpClient, HttpResult};
 
 pub struct Worker {
     http_client: &'static HttpClient,
@@ -40,7 +40,13 @@ impl Worker {
     /// will perfor delay between repeats and will check the HTTP client reponse.
     ///
     /// All the responses will be checked and recorded in `WorkerResult` struct.
-    pub async fn perform_requests(&self, request: WorkerRequest) -> Box<WorkerResult> {
+    pub async fn perform_requests(&self, requests: Vec<WorkerRequest>) -> Box<WorkerResult> {
+        if requests.len() > 1 {
+            panic!("Multiple URL are not supported yet!")
+        }
+
+        let request = requests.first().unwrap();
+
         let mut result = Box::new(WorkerResult::new());
         let mut handles: tokio::task::JoinSet<HttpResult> = tokio::task::JoinSet::new();
 
@@ -52,20 +58,18 @@ impl Worker {
             }
 
             for _ in 0..self.concurrent {
-                self.http_client
-                    .resolve_request(&request)
-                    .map_or_else(
-                        |_| {
-                            log::info!("Wrong HTTP method - skip and count skipped");
+                self.http_client.resolve_request(request).map_or_else(
+                    |_| {
+                        log::info!("Wrong HTTP method - skip and count skipped");
 
-                            log::error!("Error calling URL - wrong method: '{}'", request.method);
-                            result.totals.inc_skipped();
-                        },
-                        |req| {
-                            let future = self.http_client.execute_request(req);
-                            handles.spawn(future);
-                        },
-                    );
+                        log::error!("Error calling URL - wrong method: '{}'", request.method);
+                        result.totals.inc_skipped();
+                    },
+                    |req| {
+                        let future = self.http_client.execute_request(req);
+                        handles.spawn(future);
+                    },
+                );
             }
 
             while let Some(res) = handles.join_next().await {
